+++ tests/testthat/test-webland.R
@@ -1,5 +1,7 @@
 test_that("webland", {
   skip_on_cran()
+  skip_if(Sys.getenv("REINFOLIB_API_KEY") == "",
+          "`REINFOLIB_API_KEY` is not set.")
   library(dplyr)
 
   city <- webland_city()
@@ -14,8 +16,8 @@ test_that("webland", {
   expect_s3_class(trade, "webland_trade")
 
   trade <- trade |>
-    itemise(from = "20151",
-            to = "20151",
+    itemise(year = "2015",
+            quarter = "1",
             pref_code = "01",
             city_code = "01101") |>
     collect()